Move eventbus to system folder
